chandrahoro/backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Swiss Ephemeris for astronomical calculations
pyswisseph==2.10.3.2

# VedicAstro library for KP System calculations
vedicastro==0.2.1
# Note: flatlib sidereal branch is required by vedicastro
# Install separately: pip install git+https://github.com/diliprk/flatlib.git@sidereal#egg=flatlib

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client for external APIs (GeoNames, etc.)
httpx==0.25.2
aiohttp==3.9.1

# Date and time handling
python-dateutil==2.8.2
pytz==2023.3

# PDF generation
reportlab==4.0.7

# Image processing for chart exports
Pillow==10.1.0

# Caching and session management
redis==5.0.1
aioredis==2.0.1

# AI/ML Dependencies
anthropic==0.71.0
openai==1.54.0

# Database (if needed for user data)
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0  # PostgreSQL async driver
aiomysql==0.2.0  # MySQL async driver
pymysql==1.1.0   # MySQL driver for SQLAlchemy

# Environment variables
python-dotenv==1.0.0

# Logging and monitoring
structlog==23.2.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development dependencies
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Security
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0
email-validator==2.1.0
cryptography==41.0.7